renamed updateparticipantdialog to editparticipantdialog and new dialog format to participantrequestitem
